Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/css/css-values/minmax-length-serialize-expected.txt


FAIL e.style['letter-spacing'] = "min(1px)" should set the property value assert_equals: serialization should be canonical expected "min(1px)" but got "calc(min(1px))"
FAIL e.style['letter-spacing'] = "min(1cm)" should set the property value assert_equals: serialization should be canonical expected "min(1cm)" but got "calc(min(1cm))"
FAIL e.style['letter-spacing'] = "min(1mm)" should set the property value assert_equals: serialization should be canonical expected "min(1mm)" but got "calc(min(1mm))"
FAIL e.style['letter-spacing'] = "min(1Q)" should set the property value assert_equals: serialization should be canonical expected "min(1q)" but got "calc(min(1q))"
FAIL e.style['letter-spacing'] = "min(1in)" should set the property value assert_equals: serialization should be canonical expected "min(1in)" but got "calc(min(1in))"
FAIL e.style['letter-spacing'] = "min(1pc)" should set the property value assert_equals: serialization should be canonical expected "min(1pc)" but got "calc(min(1pc))"
FAIL e.style['letter-spacing'] = "min(1pt)" should set the property value assert_equals: serialization should be canonical expected "min(1pt)" but got "calc(min(1pt))"
FAIL e.style['letter-spacing'] = "min(1em)" should set the property value assert_equals: serialization should be canonical expected "min(1em)" but got "calc(min(1em))"
FAIL e.style['letter-spacing'] = "min(1ex)" should set the property value assert_equals: serialization should be canonical expected "min(1ex)" but got "calc(min(1ex))"
FAIL e.style['letter-spacing'] = "min(1ch)" should set the property value assert_equals: serialization should be canonical expected "min(1ch)" but got "calc(min(1ch))"
FAIL e.style['letter-spacing'] = "min(1rem)" should set the property value assert_equals: serialization should be canonical expected "min(1rem)" but got "calc(min(1rem))"
FAIL e.style['letter-spacing'] = "min(1vh)" should set the property value assert_equals: serialization should be canonical expected "min(1vh)" but got "calc(min(1vh))"
FAIL e.style['letter-spacing'] = "min(1vw)" should set the property value assert_equals: serialization should be canonical expected "min(1vw)" but got "calc(min(1vw))"
FAIL e.style['letter-spacing'] = "min(1vmin)" should set the property value assert_equals: serialization should be canonical expected "min(1vmin)" but got "calc(min(1vmin))"
FAIL e.style['letter-spacing'] = "min(1vmax)" should set the property value assert_equals: serialization should be canonical expected "min(1vmax)" but got "calc(min(1vmax))"
FAIL e.style['letter-spacing'] = "max(1px)" should set the property value assert_equals: serialization should be canonical expected "max(1px)" but got "calc(max(1px))"
FAIL e.style['letter-spacing'] = "max(1cm)" should set the property value assert_equals: serialization should be canonical expected "max(1cm)" but got "calc(max(1cm))"
FAIL e.style['letter-spacing'] = "max(1mm)" should set the property value assert_equals: serialization should be canonical expected "max(1mm)" but got "calc(max(1mm))"
FAIL e.style['letter-spacing'] = "max(1Q)" should set the property value assert_equals: serialization should be canonical expected "max(1q)" but got "calc(max(1q))"
FAIL e.style['letter-spacing'] = "max(1in)" should set the property value assert_equals: serialization should be canonical expected "max(1in)" but got "calc(max(1in))"
FAIL e.style['letter-spacing'] = "max(1pc)" should set the property value assert_equals: serialization should be canonical expected "max(1pc)" but got "calc(max(1pc))"
FAIL e.style['letter-spacing'] = "max(1pt)" should set the property value assert_equals: serialization should be canonical expected "max(1pt)" but got "calc(max(1pt))"
FAIL e.style['letter-spacing'] = "max(1em)" should set the property value assert_equals: serialization should be canonical expected "max(1em)" but got "calc(max(1em))"
FAIL e.style['letter-spacing'] = "max(1ex)" should set the property value assert_equals: serialization should be canonical expected "max(1ex)" but got "calc(max(1ex))"
FAIL e.style['letter-spacing'] = "max(1ch)" should set the property value assert_equals: serialization should be canonical expected "max(1ch)" but got "calc(max(1ch))"
FAIL e.style['letter-spacing'] = "max(1rem)" should set the property value assert_equals: serialization should be canonical expected "max(1rem)" but got "calc(max(1rem))"
FAIL e.style['letter-spacing'] = "max(1vh)" should set the property value assert_equals: serialization should be canonical expected "max(1vh)" but got "calc(max(1vh))"
FAIL e.style['letter-spacing'] = "max(1vw)" should set the property value assert_equals: serialization should be canonical expected "max(1vw)" but got "calc(max(1vw))"
FAIL e.style['letter-spacing'] = "max(1vmin)" should set the property value assert_equals: serialization should be canonical expected "max(1vmin)" but got "calc(max(1vmin))"
FAIL e.style['letter-spacing'] = "max(1vmax)" should set the property value assert_equals: serialization should be canonical expected "max(1vmax)" but got "calc(max(1vmax))"
FAIL e.style['letter-spacing'] = "min(10px, 20px, 30px)" should set the property value assert_equals: serialization should be canonical expected "min(10px, 20px, 30px)" but got "calc(min(10px, 20px, 30px))"
FAIL e.style['letter-spacing'] = "min(30px, 20px, 10px)" should set the property value assert_equals: serialization should be canonical expected "min(30px, 20px, 10px)" but got "calc(min(30px, 20px, 10px))"
FAIL e.style['letter-spacing'] = "min(20px, 1em, 10vw)" should set the property value assert_equals: serialization should be canonical expected "min(20px, 1em, 10vw)" but got "calc(min(20px, 1em, 10vw))"
FAIL e.style['letter-spacing'] = "min(10vw, 1em, 20px)" should set the property value assert_equals: serialization should be canonical expected "min(10vw, 1em, 20px)" but got "calc(min(10vw, 1em, 20px))"
FAIL e.style['letter-spacing'] = "max(10px, 20px, 30px)" should set the property value assert_equals: serialization should be canonical expected "max(10px, 20px, 30px)" but got "calc(max(10px, 20px, 30px))"
FAIL e.style['letter-spacing'] = "max(30px, 20px, 10px)" should set the property value assert_equals: serialization should be canonical expected "max(30px, 20px, 10px)" but got "calc(max(30px, 20px, 10px))"
FAIL e.style['letter-spacing'] = "max(20px, 1em, 10vw)" should set the property value assert_equals: serialization should be canonical expected "max(20px, 1em, 10vw)" but got "calc(max(20px, 1em, 10vw))"
FAIL e.style['letter-spacing'] = "max(10vw, 1em, 20px)" should set the property value assert_equals: serialization should be canonical expected "max(10vw, 1em, 20px)" but got "calc(max(10vw, 1em, 20px))"
FAIL e.style['letter-spacing'] = "calc(min(10px) + max(1em) + min(10vw))" should set the property value assert_equals: serialization should be canonical expected "calc(min(10px) + max(1em) + min(10vw))" but got "calc(((min(10px)) + (max(1em))) + (min(10vw)))"
FAIL e.style['letter-spacing'] = "calc(max(1em) + min(10vw) + max(10px))" should set the property value assert_equals: serialization should be canonical expected "calc(max(1em) + min(10vw) + max(10px))" but got "calc(((max(1em)) + (min(10vw))) + (max(10px)))"
FAIL e.style['letter-spacing'] = "calc(10px + min(1em))" should set the property value assert_equals: serialization should be canonical expected "calc(10px + min(1em))" but got "calc(10px + (min(1em)))"
FAIL e.style['letter-spacing'] = "calc(min(1em) + 10px)" should set the property value assert_equals: serialization should be canonical expected "calc(10px + min(1em))" but got "calc((min(1em)) + 10px)"
FAIL e.style['letter-spacing'] = "calc(10px + max(1em))" should set the property value assert_equals: serialization should be canonical expected "calc(10px + max(1em))" but got "calc(10px + (max(1em)))"
FAIL e.style['letter-spacing'] = "calc(max(1em) + 10px)" should set the property value assert_equals: serialization should be canonical expected "calc(10px + max(1em))" but got "calc((max(1em)) + 10px)"
